# openapi/v1.yaml
...
openapi: "3.0.0"
info:
  version: "1.0.0-alpha1"
  description: "API definition for Genexec, a generic execution platform for IaC"
  title: "Genexec OpenAPI"
  contact:
    name: "Genexec"
    email: "gexec@webhippie.de"
    url: "https://gexec.eu"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://try.gexec.eu/api/v1"
security:
  - Header: []
  - Bearer: []
  - Basic: []
paths:
  /auth/{provider}/request:
    get:
      summary: "Request the redirect to defined provider"
      operationId: "RequestProvider"
      tags:
        - "auth"
      parameters:
        - $ref: "#/components/parameters/AuthProviderParam"
      security: []
      responses:
        "308":
          $ref: "#/components/responses/ProviderRequestResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /auth/{provider}/callback:
    get:
      summary: "Callback to parse the defined provider"
      operationId: "CallbackProvider"
      tags:
        - "auth"
      parameters:
        - $ref: "#/components/parameters/AuthProviderParam"
        - $ref: "#/components/parameters/AuthStateParam"
        - $ref: "#/components/parameters/AuthCodeParam"
      security: []
      responses:
        "308":
          $ref: "#/components/responses/ProviderCallbackResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "412":
          $ref: "#/components/responses/ProviderInitError"
...
  /auth/login:
    post:
      summary: "Authenticate an user by credentials"
      operationId: "LoginAuth"
      tags:
        - "auth"
      requestBody:
        $ref: "#/components/requestBodies/LoginAuthBody"
      security: []
      responses:
        "200":
          $ref: "#/components/responses/LoginResponse"
        "401":
          $ref: "#/components/responses/BadCredentialsError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/refresh:
    get:
      summary: "Refresh an auth token before it expires"
      operationId: "RefreshAuth"
      tags:
        - "auth"
      responses:
        "200":
          $ref: "#/components/responses/RefreshResponse"
        "401":
          $ref: "#/components/responses/InvalidTokenError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/verify:
    get:
      summary: "Verify validity for an authentication token"
      operationId: "VerifyAuth"
      tags:
        - "auth"
      responses:
        "200":
          $ref: "#/components/responses/VerifyResponse"
        "401":
          $ref: "#/components/responses/InvalidTokenError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /profile/token:
    get:
      summary: "Retrieve an unlimited auth token"
      operationId: "TokenProfile"
      tags:
        - "profile"
      responses:
        "200":
          $ref: "#/components/responses/TokenResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /profile/self:
    get:
      summary: "Fetch profile details of the personal account"
      operationId: "ShowProfile"
      tags:
        - "profile"
      responses:
        "200":
          $ref: "#/components/responses/ProfileResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: "Update your own profile information"
      operationId: "UpdateProfile"
      tags:
        - "profile"
      requestBody:
        $ref: "#/components/requestBodies/UpdateProfileBody"
      responses:
        "200":
          $ref: "#/components/responses/ProfileResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects:
    get:
      summary: "Fetch all available projects"
      operationId: "ListProjects"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/SearchQueryParam"
        - $ref: "#/components/parameters/SortColumnParam"
        - $ref: "#/components/parameters/SortOrderParam"
        - $ref: "#/components/parameters/PagingLimitParam"
        - $ref: "#/components/parameters/PagingOffsetParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectsResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: "Create a new project"
      operationId: "CreateProject"
      tags:
        - "project"
      requestBody:
        $ref: "#/components/requestBodies/CreateProjectBody"
      responses:
        "200":
          $ref: "#/components/responses/ProjectResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{project_id}:
    get:
      summary: "Fetch a specific project"
      operationId: "ShowProject"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: "Update a specific project"
      operationId: "UpdateProject"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      requestBody:
        $ref: "#/components/requestBodies/UpdateProjectBody"
      responses:
        "200":
          $ref: "#/components/responses/ProjectResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: "Delete a specific project"
      operationId: "DeleteProject"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      responses:
        "200":
          $ref: "#/components/responses/SuccessMessage"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/ActionFailedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{project_id}/users:
    get:
      summary: "Fetch all users attached to project"
      operationId: "ListProjectUsers"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/SearchQueryParam"
        - $ref: "#/components/parameters/SortColumnParam"
        - $ref: "#/components/parameters/SortOrderParam"
        - $ref: "#/components/parameters/PagingLimitParam"
        - $ref: "#/components/parameters/PagingOffsetParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectUsersResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: "Attach a user to project"
      operationId: "AttachProjectToUser"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      requestBody:
        $ref: "#/components/requestBodies/ProjectUserPermBody"
      responses:
        "200":
          $ref: "#/components/responses/SuccessMessage"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "412":
          $ref: "#/components/responses/AlreadyAttachedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: "Update user perms for project"
      operationId: "PermitProjectUser"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      requestBody:
        $ref: "#/components/requestBodies/ProjectUserPermBody"
      responses:
        "200":
          $ref: "#/components/responses/SuccessMessage"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "412":
          $ref: "#/components/responses/NotAttachedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: "Unlink a user from project"
      operationId: "DeleteProjectFromUser"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      requestBody:
        $ref: "#/components/requestBodies/ProjectUserDropBody"
      responses:
        "200":
          $ref: "#/components/responses/SuccessMessage"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "412":
          $ref: "#/components/responses/NotAttachedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{project_id}/groups:
    get:
      summary: "Fetch all groups attached to project"
      operationId: "ListProjectGroups"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/SearchQueryParam"
        - $ref: "#/components/parameters/SortColumnParam"
        - $ref: "#/components/parameters/SortOrderParam"
        - $ref: "#/components/parameters/PagingLimitParam"
        - $ref: "#/components/parameters/PagingOffsetParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectGroupsResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: "Attach a group to project"
      operationId: "AttachProjectToGroup"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      requestBody:
        $ref: "#/components/requestBodies/ProjectGroupPermBody"
      responses:
        "200":
          $ref: "#/components/responses/SuccessMessage"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "412":
          $ref: "#/components/responses/AlreadyAttachedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: "Update group perms for project"
      operationId: "PermitProjectGroup"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      requestBody:
        $ref: "#/components/requestBodies/ProjectGroupPermBody"
      responses:
        "200":
          $ref: "#/components/responses/SuccessMessage"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "412":
          $ref: "#/components/responses/NotAttachedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: "Unlink a group from project"
      operationId: "DeleteProjectFromGroup"
      tags:
        - "project"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      requestBody:
        $ref: "#/components/requestBodies/ProjectGroupDropBody"
      responses:
        "200":
          $ref: "#/components/responses/SuccessMessage"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "412":
          $ref: "#/components/responses/NotAttachedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{project_id}/events:
    get:
      summary: "Fetch all events for a project"
      operationId: "ListProjectEvents"
      tags:
        - "event"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/SearchQueryParam"
        - $ref: "#/components/parameters/PagingLimitParam"
        - $ref: "#/components/parameters/PagingOffsetParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectEventsResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{project_id}/runners:
    get:
      summary: "Fetch all runners for a project"
      operationId: "ListProjectRunners"
      tags:
        - "project"
        - "runner"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/SearchQueryParam"
        - $ref: "#/components/parameters/SortColumnParam"
        - $ref: "#/components/parameters/SortOrderParam"
        - $ref: "#/components/parameters/PagingLimitParam"
        - $ref: "#/components/parameters/PagingOffsetParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectRunnersResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      summary: "Create a new runner"
      operationId: "CreateProjectRunner"
      tags:
        - "project"
        - "runner"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
      requestBody:
        $ref: "#/components/requestBodies/CreateProjectRunnerBody"
      responses:
        "200":
          $ref: "#/components/responses/ProjectRunnerResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /projects/{project_id}/runners/{runner_id}:
    get:
      summary: "Fetch a specific runner for a project"
      operationId: "ShowProjectRunner"
      tags:
        - "project"
        - "runner"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/RunnerParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectRunnerResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      summary: "Update a specific runner for a project"
      operationId: "UpdateProjectRunner"
      tags:
        - "project"
        - "runner"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/RunnerParam"
      requestBody:
        $ref: "#/components/requestBodies/UpdateProjectRunnerBody"
      responses:
        "200":
          $ref: "#/components/responses/ProjectRunnerResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: "Delete a specific runner for a project"
      operationId: "DeleteProjectRunner"
      tags:
        - "project"
        - "runner"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/RunnerParam"
      responses:
        "200":
          $ref: "#/components/responses/SuccessMessage"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/ActionFailedError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /projects/{project_id}/environments/{environment_id}/secrets:
    post:
      summary: "Create a new secret on an environment"
      operationId: "CreateProjectEnvironmentSecret"
      tags:
        - "project"
        - "environment"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/EnvironmentParam"
      requestBody:
        $ref: "#/components/requestBodies/CreateProjectEnvironmentSecretBody"
      responses:
        "200":
          $ref: "#/components/responses/ProjectEnvironmentSecretResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /projects/{project_id}/executions/{execution_id}/output:
    get:
      summary: "Output a specific execution for a project"
      operationId: "OutputProjectExecution"
      tags:
        - "project"
        - "execution"
      parameters:
        - $ref: "#/components/parameters/ProjectParam"
        - $ref: "#/components/parameters/ExecutionParam"
      responses:
        "200":
          $ref: "#/components/responses/ProjectOutputResponse"
        "403":
          $ref: "#/components/responses/NotAuthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    Header:
      type: apiKey
      in: header
      name: X-API-Key
    Bearer:
      type: http
      scheme: bearer
    Basic:
      type: http
      scheme: basic
  parameters:
...
    AuthProviderParam:
      in: "path"
      name: "provider"
      description: "An identifier for the auth provider"
      schema:
        type: "string"
      required: true
      x-example: "github"
    AuthStateParam:
      name: "state"
      in: "query"
      required: false
      schema:
        type: "string"
      description: "Auth state"
    AuthCodeParam:
      name: "code"
      in: "query"
      required: false
      schema:
        type: "string"
      description: "Auth code"
...
  responses:
    NotFoundError:
      description: "Resource not found"
      content:
        text/html:
          schema:
            type: "string"
        application/json:
          schema:
            $ref: "#/components/schemas/Notification"
...
    ProviderRequestResponse:
      description: "Redirect to the defined provider"
      content:
        text/html:
          schema:
            type: "string"
    ProviderCallbackResponse:
      description: "Generated token with expire date"
      content:
        text/html:
          schema:
            type: "string"
...